+++ applications/fishing-map/src/features/map/Map.tsx
@@ -20,6 +20,7 @@ import {
 import {
   ExtendedStyle,
   ExtendedStyleMeta,
+  Generators,
   LayerMetadataLegend,
 } from '@globalfishingwatch/layer-composer'
 import { UrlDataviewInstance } from 'types'
@@ -80,38 +81,52 @@ const getLegendLayers = (
       const id = sublayerLegendMetadata.id || (layer.metadata?.generatorId as string)
       const dataview = dataviews?.find((d) => d.id === id)
       const isSquareKm = (sublayerLegendMetadata.gridArea as number) > 50000
-      const gridArea = isSquareKm
-        ? (sublayerLegendMetadata.gridArea as number) / 1000000
-        : sublayerLegendMetadata.gridArea
-      const gridAreaFormatted = gridArea
-        ? formatI18nNumber(gridArea, {
-            style: 'unit',
-            unit: isSquareKm ? 'kilometer' : 'meter',
-            unitDisplay: 'short',
-          })
-        : ''
+      let label = sublayerLegendMetadata.unit
+      if (!label) {
+        const gridArea = isSquareKm
+          ? (sublayerLegendMetadata.gridArea as number) / 1000000
+          : sublayerLegendMetadata.gridArea
+        const gridAreaFormatted = gridArea
+          ? formatI18nNumber(gridArea, {
+              style: 'unit',
+              unit: isSquareKm ? 'kilometer' : 'meter',
+              unitDisplay: 'short',
+            })
+          : ''
+        label = `${i18n.t('common.hour_plural', 'hours')} / ${gridAreaFormatted}`
+      }
       const sublayerLegend: LegendLayer | LegendLayerBivariate = {
         ...sublayerLegendMetadata,
         id: `legend_${id}`,
         color: layer.metadata?.color || dataview?.config?.color || 'red',
-        label: `${i18n.t('common.hour_plural', 'hours')} / ${gridAreaFormatted}`,
+        label,
       }
 
-      const getHoveredFeatureValueForSublayerIndex = (index: number): number => {
-        const hoveredFeature = hoveredEvent?.features?.find(
-          (f) => f.temporalgrid?.sublayerIndex === index
-        )
-        return hoveredFeature?.value
-      }
+      const generatorType = layer.metadata?.generatorType
 
-      // Both bivariate sublayers come in the same sublayerLegend (see getLegendsBivariate in LC)
-      if (sublayerLegend.type === 'bivariate') {
-        sublayerLegend.currentValues = [
-          getHoveredFeatureValueForSublayerIndex(0),
-          getHoveredFeatureValueForSublayerIndex(1),
-        ]
-      } else {
-        sublayerLegend.currentValue = getHoveredFeatureValueForSublayerIndex(sublayerIndex)
+      if (generatorType === Generators.Type.Heatmap) {
+        const value = hoveredEvent?.features?.find(
+          (f) => f.generatorId === layer.metadata?.generatorId
+        )?.value
+        if (value) {
+          sublayerLegend.currentValue = value
+        }
+      } else if (generatorType === Generators.Type.HeatmapAnimated) {
+        const getHoveredFeatureValueForSublayerIndex = (index: number): number => {
+          const hoveredFeature = hoveredEvent?.features?.find(
+            (f) => f.temporalgrid?.sublayerIndex === index
+          )
+          return hoveredFeature?.value
+        }
+        // Both bivariate sublayers come in the same sublayerLegend (see getLegendsBivariate in LC)
+        if (sublayerLegend.type === 'bivariate') {
+          sublayerLegend.currentValues = [
+            getHoveredFeatureValueForSublayerIndex(0),
+            getHoveredFeatureValueForSublayerIndex(1),
+          ]
+        } else {
+          sublayerLegend.currentValue = getHoveredFeatureValueForSublayerIndex(sublayerIndex)
+        }
       }
       return sublayerLegend
     })
@@ -165,6 +180,9 @@ const MapWrapper = (): React.ReactElement | null => {
     style?.metadata
   )
   const hoveredTooltipEvent = useMapTooltip(hoveredEvent)
+  const onMouseOut = useCallback(() => {
+    setHoveredEvent(null)
+  }, [])
 
   const { viewport, onViewportChange } = useViewport()
 
@@ -178,6 +196,7 @@ const MapWrapper = (): React.ReactElement | null => {
   const layersWithLegend = getLegendLayers(style, dataviews, hoveredEvent)
 
   const debugOptions = useSelector(selectDebugOptions)
+
   const getRulersCursor = useCallback(() => {
     return 'crosshair'
   }, [])
@@ -224,6 +243,7 @@ const MapWrapper = (): React.ReactElement | null => {
           onClick={onMapClick}
           onHover={onMapHover}
           onError={handleError}
+          onMouseOut={onMouseOut}
           transitionDuration={viewport.transitionDuration}
         >
           {clickedEvent && (
@@ -255,7 +275,7 @@ const MapWrapper = (): React.ReactElement | null => {
               labelComponent={
                 <span className={styles.legendLabel}>
                   {legend.label}
-                  <sup>2</sup>
+                  {legend.gridArea && <sup>2</sup>}
                 </span>
               }
             />,