+++ applications/fishing-map/src/features/map/popups/PopupWrapper.tsx
@@ -7,6 +7,7 @@ import { Generators } from '@globalfishingwatch/layer-composer'
 import { TooltipEvent } from 'features/map/map.hooks'
 import styles from './Popup.module.css'
 import HeatmapTooltipRow from './HeatmapLayers'
+import EnviromentalTooltipSection from './EnvironmentLayers'
 import ContextTooltipSection from './ContextLayers'
 
 type PopupWrapperProps = {
@@ -57,6 +58,15 @@ function PopupWrapper({
           if (featureType === Generators.Type.Context) {
             return (
               <ContextTooltipSection
+                key={featureType}
+                features={features}
+                showFeaturesDetails={type === 'click'}
+              />
+            )
+          }
+          if (featureType === Generators.Type.Heatmap) {
+            return (
+              <EnviromentalTooltipSection
                 key={featureType}
                 features={features}
                 showFeaturesDetails={type === 'click'}