+++ applications/fishing-map/src/features/user/User.tsx
@@ -12,7 +12,6 @@ import {
 } from 'features/workspaces-list/workspaces-list.slice'
 import { WORKSPACE } from 'routes/routes'
 import { WorkspaceCategories } from 'data/workspaces'
-import { APP_NAME } from 'data/config'
 import styles from './User.module.css'
 import { logoutUserThunk, selectUserData } from './user.slice'
 import { isUserLogged, selectUserWorkspaces } from './user.selectors'
@@ -27,8 +26,8 @@ function User() {
   const [logoutLoading, setLogoutLoading] = useState(false)
 
   useEffect(() => {
-    if (userLogged && workspacesStatus !== AsyncReducerStatus.Finished) {
-      dispatch(fetchWorkspacesThunk(APP_NAME))
+    if (userLogged) {
+      dispatch(fetchWorkspacesThunk({ userId: userData?.id }))
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
@@ -74,8 +73,7 @@ function User() {
                     to={{
                       type: WORKSPACE,
                       payload: {
-                        // TODO change to workspace category (save in API)
-                        category: WorkspaceCategories.FishingActivity,
+                        category: workspace.category || WorkspaceCategories.FishingActivity,
                         workspaceId: workspace.id,
                       },
                       query: {},