+++ packages/react-hooks/src/use-map-interaction/use-map-interaction.ts
@@ -50,16 +50,17 @@ const aggregateCell = (
 
 const getExtendedFeatures = (
   features: MapboxGeoJSONFeature[],
-  metatada?: ExtendedStyleMeta
+  metadata?: ExtendedStyleMeta
 ): ExtendedFeature[] => {
-  const timeChunks = metatada?.temporalgrid?.timeChunks
+  const timeChunks = metadata?.temporalgrid?.timeChunks
   const frame = timeChunks?.activeChunkFrame
   const activeTimeChunk = timeChunks?.chunks.find((c: any) => c.active)
-  const numSublayers = metatada?.temporalgrid?.numSublayers
+  const numSublayers = metadata?.temporalgrid?.numSublayers
 
   const extendedFeatures: ExtendedFeature[] = features.flatMap((feature: MapboxGeoJSONFeature) => {
     const generatorType = feature.layer.metadata?.generatorType ?? null
     const generatorId = feature.layer.metadata?.generatorId ?? null
+    const unit = feature.layer?.metadata?.legend?.unit ?? null
     const properties = feature.properties || {}
     const extendedFeature: ExtendedFeature | null = {
       properties,
@@ -69,6 +70,7 @@ const getExtendedFeatures = (
       sourceLayer: feature.sourceLayer,
       id: (feature.id as number) || feature.properties?.gfw_id || undefined,
       value: properties.value || properties.name || properties.id,
+      unit,
       tile: {
         x: (feature as any)._vectorTileFeature._x,
         y: (feature as any)._vectorTileFeature._y,