+++ applications/fishing-map/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @globalfishingwatchapp/fishing-map
 
+## 0.17.5
+
+### Patch Changes
+
+- 4f8269c2: [WIP] Polish 🇵🇱
+- Updated dependencies [4f8269c2]
+  - @globalfishingwatch/ui-components@2.11.3
+
+## 0.17.4
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@4.11.0
+
+## 0.17.3
+
+### Patch Changes
+
+- d160e84a: update dependencies
+- Updated dependencies [d160e84a]
+  - @globalfishingwatch/layer-composer@4.10.4
+
 ## 0.17.2
 
 ### Patch Changes